1. 장신구 옵션 종류 6 -> 9개 수정 2. 도전의 탑 층수 초기화 유저 개별로 하도록 수정 3. 월드 보스 점수 무한대로 수정

+++ IdleGameServer/src/main/java/com/onlyonegames/eternalfantasia/domain/model/dto/Contents/PreviousWorldBossRankingDto.java
@@ -11,8 +11,8 @@ public class PreviousWorldBossRankingDto {
     Long id;
     Long useridUser;
     String userGameName;
-    Long totalDamage;
-    Long bestDamage;
+    double totalDamage;
+    double bestDamage;
     int ranking;
     boolean receivable;
